remove old call playback class
